Extract classes in separate files
